Implemented tutorial puzzle
Added and placed the bridge and NPC sprites.
Made a prefab of the NPC.
Added the coding UI and made the tutorial puzzle (tell the NPC your name).
Added a chest with the print block.
Added an inventory to Hedy with a block 'I'm Hedy'.
Added the eventsystem because the coding UI doesn't work without it.